+++ spec/modules/recon/x_forwarded_for_access_restriction_bypass_spec.rb
@@ -8,11 +8,11 @@ describe name_from_filename do
     end
 
     def self.elements
-        [ Element::FORM, Element::LINK, Element::COOKIE, Element::HEADER ]
+        [ Element::SERVER ]
     end
 
     def issue_count
-        12
+        2
     end
 
     easy_test