+++ spec/support/servers/modules/recon/grep/form_upload.rb
@@ -1,0 +1,9 @@
+require 'sinatra'
+
+get '/' do
+    <<-EOHTML
+        <form name="upload-form">
+            <input type="file" name="upload-me">
+        </form>
+    EOHTML
+end